services/api/requirements.txt
# Core Django and web framework
Django==4.2.23
django-cors-headers==4.3.1

# Authentication and permissions
django-allauth==0.54.0
dj-rest-auth==4.0.1
PyJWT==2.8.0

# API and serialization
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2

# Database
# Upgrade to a wheel that supports Python 3.12 on Windows
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Cache
redis==5.0.1
django-redis==5.4.0

# File storage (AWS S3)
boto3==1.34.0
django-storages==1.14.2

# Security
django-csp==3.8
whitenoise==6.7.0
gunicorn==21.2.0

# Environment management
python-decouple==3.8
python-dotenv==1.0.1

# Filtering utilities used by 'django_filters' app
django-filter==24.2

# Payments
stripe==6.4.0

# External integrations
google-generativeai==0.5.0
requests==2.31.0

# Validation and utilities
Pillow==10.0.0

# Development tools (remove in production if needed)
django-debug-toolbar==4.2.0